requirements.txt
# Core LLM Libraries
llama-parse
llama-index
llama-index-core
llama-index-llms-openai
openai

# Google APIs
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-auth

# Data Processing
pandas
openpyxl
pydantic

# Utilities
python-dotenv
asyncio-throttle

# Development & Testing
pytest
pytest-asyncio
